refactor: rename custom fasthttp dialer/resolver
